requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
distro==1.9.0
docarray==0.40.0
Flask==3.0.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-openai==0.1.1
langchain-pinecone==0.0.3
langchain-text-splitters==0.0.1
langsmith==0.1.31
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.14.3
orjson==3.9.15
packaging==23.2
pinecone-client==3.2.1
pydantic==1.10.8
Pygments==2.17.2
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
ruff==0.3.4
scikit-learn==1.4.1.post1
scipy==1.12.0
sniffio==1.3.1
SQLAlchemy==2.0.29
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tqdm==4.66.2
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
